pyproject.toml
[tool.poetry]
name = "contacts-api"
version = "0.1.0"
description = "REST API for managing contacts using FastAPI and SQLAlchemy"
authors = ["Roman Kukla <kukla.roman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.7"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
email-validator = "^2.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
slowapi = "^0.1.9"
cloudinary = "^1.36.0"
jinja2 = "^3.1.2"
aiofiles = "^23.2.1"
fastapi-mail = "^1.4.1"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 